+++ alicloud/data_source_alicloud_zones.go
@@ -6,7 +6,9 @@ import (
 	"sort"
 	"strings"
 
+	"github.com/aliyun/alibaba-cloud-sdk-go/services/r-kvstore"
 	"github.com/aliyun/alibaba-cloud-sdk-go/services/rds"
+
 	//"github.com/denverdino/aliyungo/ecs"
 	"github.com/aliyun/alibaba-cloud-sdk-go/services/ecs"
 	"github.com/hashicorp/terraform/helper/schema"
@@ -30,6 +32,7 @@ func dataSourceAlicloudZones() *schema.Resource {
 				ValidateFunc: validateAllowedStringValue([]string{
 					string(ResourceTypeInstance),
 					string(ResourceTypeRds),
+					string(ResourceTypeRkv),
 					string(ResourceTypeVSwitch),
 					string(ResourceTypeDisk),
 					string(IoOptimized),
@@ -114,6 +117,7 @@ func dataSourceAlicloudZonesRead(d *schema.ResourceData, meta interface{}) error
 	client := meta.(*AliyunClient)
 	var zoneIds []string
 	rdsZones := make(map[string]string)
+	rkvZones := make(map[string]string)
 	if strings.ToLower(Trim(resType)) == strings.ToLower(string(ResourceTypeRds)) {
 		request := rds.CreateDescribeRegionsRequest()
 		if regions, err := client.rdsconn.DescribeRegions(request); err != nil {
@@ -127,6 +131,24 @@ func dataSourceAlicloudZonesRead(d *schema.ResourceData, meta interface{}) error
 					continue
 				}
 				rdsZones[r.ZoneId] = r.RegionId
+			}
+		}
+	}
+	if strings.ToLower(Trim(resType)) == strings.ToLower(string(ResourceTypeRkv)) {
+		request := r_kvstore.CreateDescribeRegionsRequest()
+		if regions, err := client.rkvconn.DescribeRegions(request); err != nil {
+			return fmt.Errorf("[ERROR] DescribeRegions got an error: %#v", err)
+		} else if len(regions.RegionIds.KVStoreRegion) <= 0 {
+			return fmt.Errorf("[ERROR] There is no available region for KVStore")
+		} else {
+			for _, r := range regions.RegionIds.KVStoreRegion {
+				for _, zoneID := range strings.Split(r.ZoneIds, ",") {
+					if multi && strings.Contains(zoneID, MULTI_IZ_SYMBOL) && r.RegionId == string(getRegion(d, meta)) {
+						zoneIds = append(zoneIds, zoneID)
+						continue
+					}
+					rkvZones[zoneID] = r.RegionId
+				}
 			}
 		}
 	}
@@ -158,6 +180,7 @@ func dataSourceAlicloudZonesRead(d *schema.ResourceData, meta interface{}) error
 	if resp == nil || len(resp.Zones.Zone) < 1 {
 		return fmt.Errorf("There are no availability zones in the region: %#v.", getRegion(d, meta))
 	}
+
 	mapZones := make(map[string]ecs.Zone)
 	insType, _ := d.Get("available_instance_type").(string)
 	diskType, _ := d.Get("available_disk_category").(string)
@@ -177,6 +200,11 @@ func dataSourceAlicloudZonesRead(d *schema.ResourceData, meta interface{}) error
 			}
 			if len(rdsZones) > 0 {
 				if _, ok := rdsZones[zone.ZoneId]; !ok {
+					continue
+				}
+			}
+			if len(rkvZones) > 0 {
+				if _, ok := rkvZones[zone.ZoneId]; !ok {
 					continue
 				}
 			}