+++ alicloud/diff_suppress_funcs.go
@@ -240,3 +240,10 @@ func routerInterfaceVBRTypeDiffSuppressFunc(k, old, new string, d *schema.Resour
 	}
 	return false
 }
+
+func rkvPostPaidDiffSuppressFunc(k, old, new string, d *schema.ResourceData) bool {
+	if rds.DBPayType(d.Get("instance_charge_type").(string)) == rds.Prepaid {
+		return false
+	}
+	return true
+}